# Conflicts:
#	.idea/caches/build_file_checksums.ser
#	.idea/caches/gradle_models.ser
#	app/src/main/java/com/operatorsapp/fragments/SelectMachineFragment.java
#	app/src/main/res/layout/fragment_select_machine.xml
#	app/src/main/res/values/strings.xml
#	oppapplog/build/intermediates/incremental/packageDebugResources/merger.xml
#	oppapplog/build/outputs/logs/manifest-merger-debug-report.txt

+++ app/src/main/java/com/operatorsapp/fragments/ReportCycleUnitsFragment.java
@@ -72,7 +72,7 @@ public class ReportCycleUnitsFragment extends BackStackAwareFragment implements 
     private double mUnitsCounter = 1;
     private ReportCore mReportCore;
     private Integer mJobId = null;
-    private int mMaxUnits = 0;
+    private float mMaxUnits = 0f;
 
     private ActiveJobsListForMachine mActiveJobsListForMachine;
     private Spinner mJobsSpinner;
